Give up on Grunt and Compass
Bower's requirement that build artifacts be in the repository makes me
too sad to bother with a build process. I will instead write a sad and
lonely js file. Its only friend will be an equally sad css file that
knows not the beauty of being sassy or the knowledge of how to use a
compass.